Move mockData file
